Issue: beacon accelerometer init order #48
The beacon `BeaconAccelHelper` is initialized at the `klippy:connect` stage, so if motors_sync registers a connect callback before the beacon, the object will not exist at the connect stage for motors_sync. The order of registration determines the tab order in printer.cfg. So now on the first connect callback we register another callback to the end of the queue.